+++ .gitignore
@@ -6,3 +6,5 @@ target
 perf.data*
 
 Cargo.lock
+
+lindera-ipadic-*.tar.bz2
