+++ Cargo.toml
@@ -1,9 +1,9 @@
 [package]
 name = "lindera"
-version = "0.1.4"
+version = "0.1.5"
 authors = ["Minoru Osuka <minoru.osuka@gmail.com>"]
 edition = "2018"
-description = "A Japanese Morphological Analysis Library written in Rust."
+description = "A Japanese Morphological Analysis Library."
 homepage = "https://github.com/bayard-search/lindera"
 repository = "https://github.com/bayard-search/lindera"
 readme = "README.md"
@@ -15,7 +15,6 @@ license = "MIT"
 bincode = "1.2.1"
 byteorder = "1.3.2"
 encoding = "0.2.33"
-lindera-ipadic = "0.1.3"
 serde = {version="1.0.104", features = ["derive"] }
 tantivy-fst = "0.2.1"
 
@@ -23,10 +22,8 @@ tantivy-fst = "0.2.1"
 criterion = "0.3.1"
 
 [profile.release]
-debug = false
 lto = true
 
 [[bench]]
 name = "bench"
 harness = false
-
