--- src/ipadic.rs
@@ -1,9 +1,0 @@
-pub mod builder;
-pub mod character_definition;
-pub mod connection;
-pub mod prefix_dict;
-pub mod system_dict;
-pub mod tokenizer;
-pub mod unknown_dictionary;
-pub mod viterbi;
-pub mod word_entry;
